+++ _RELEASE/CHANGELOG.md
@@ -4,5 +4,10 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.2.0]
+ - Added chat commands ``/fov`` and ``/cfov`` to change fov/crouch fov in-game
+ - Added crouch fov config item ``CrouchFov`` for when the player is crouched (tumble mode)
+ - Hopefully fixed issue of fov not changing on respawn by adding a spawn patch
+
 ## [0.1.0]
  - Initial Release.